Add compatability changes
The old version of the game depended on some features which had not yet
been implemented in the base Stacko interpreter. This commit implements
those features in Stacko to avoid depending on a modified interpreter.
These features can be removed from the codebase as they are added into
the base interpreter.